refactor fixture
