+++ sharding-core/sharding-core-rewrite/src/test/java/org/apache/shardingsphere/core/rewrite/fixture/NormalEncryptorFixture.java
@@ -19,23 +19,32 @@ package org.apache.shardingsphere.core.rewrite.fixture;
 
 import lombok.Getter;
 import lombok.Setter;
-import org.apache.shardingsphere.spi.keygen.ShardingKeyGenerator;
+import org.apache.shardingsphere.spi.encrypt.ShardingEncryptor;
 
 import java.util.Properties;
 
 @Getter
 @Setter
-public final class TestShardingKeyGenerator implements ShardingKeyGenerator {
+public final class NormalEncryptorFixture implements ShardingEncryptor {
     
     private Properties properties = new Properties();
     
     @Override
-    public Comparable<?> generateKey() {
-        return 1L;
+    public String getType() {
+        return "NORMAL_ENCRYPT";
     }
     
     @Override
-    public String getType() {
-        return "TEST";
+    public void init() {
+    }
+    
+    @Override
+    public String encrypt(final Object plaintext) {
+        return "encrypt_" + plaintext;
+    }
+    
+    @Override
+    public Object decrypt(final String ciphertext) {
+        return ciphertext.replaceAll("encrypt_", "");
     }
 }
