+++ sharding-core/sharding-core-rewrite/src/test/java/org/apache/shardingsphere/core/rewrite/fixture/ShardingKeyGeneratorFixture.java
@@ -19,32 +19,23 @@ package org.apache.shardingsphere.core.rewrite.fixture;
 
 import lombok.Getter;
 import lombok.Setter;
-import org.apache.shardingsphere.spi.encrypt.ShardingEncryptor;
+import org.apache.shardingsphere.spi.keygen.ShardingKeyGenerator;
 
 import java.util.Properties;
 
 @Getter
 @Setter
-public final class TestShardingEncryptor implements ShardingEncryptor {
+public final class ShardingKeyGeneratorFixture implements ShardingKeyGenerator {
     
     private Properties properties = new Properties();
     
     @Override
+    public Comparable<?> generateKey() {
+        return 1L;
+    }
+    
+    @Override
     public String getType() {
         return "TEST";
-    }
-    
-    @Override
-    public void init() {
-    }
-    
-    @Override
-    public String encrypt(final Object plaintext) {
-        return "encrypt_" + plaintext;
-    }
-    
-    @Override
-    public Object decrypt(final String ciphertext) {
-        return ciphertext.replaceAll("encrypt_", "");
     }
 }
