+++ sharding-core/sharding-core-rewrite/src/test/resources/META-INF/services/org.apache.shardingsphere.spi.keygen.ShardingKeyGenerator
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.core.rewrite.fixture.TestShardingKeyGenerator
+org.apache.shardingsphere.core.rewrite.fixture.ShardingKeyGeneratorFixture
